🐛 Fix build errors: WelcomeDayContent type, OSChatScreen const, Linux STT compatibility
- Fix WelcomeDayContent → WelcomeDay type references
- Remove const from OSChatScreen navigation (not a const constructor)
- Fix os_metrics_service double return type
- Add platform check for STT (disable on Linux due to record package compatibility)
- Update welcome series to use WelcomeDay from content file

+++ lib/services/os_metrics_service.dart
@@ -123,7 +123,7 @@ class OSMetricsService {
     // Perfect consistency (0 variance) = 100, high variance (>30) = 0
     final consistencyScore = (100 - (stdDev / 30 * 100)).clamp(0, 100);
     
-    return consistencyScore;
+    return consistencyScore.toDouble();
   }
 
   /// 🎨 GET DISCIPLINE COLOR
